Make sure fontaine--set-faces does not mix up faces
The previous design was applying the styles of the 'default' face
correctly, but everything else was not done correctly. I noticed this
problem while working with buffers that relied on 'variable-pitch' and
'fixed-pitch'.

Also thanks to Haruko for reporting what essentially is the same
underlying problem in issue 14: <https://github.com/protesilaos/fontaine/issues/14>.